Update main.swift
Just putting in code to test out later. It won't work properly, and a lot of this code will not be in the main.swift file, it's just currently there for testing purposes. I will also try to advance a little past the MVP of the PenisParser, but that will come much later.